Mesh working

+++ src/TimerMenu.cpp
@@ -15,13 +15,13 @@ void TimerMenu::displayMenu(){
     colonShown = false;
     break;
   case INTERVAL1:
-    sprintf(displayText,"1 %02d%02d", menuOptions[activeMenu].getStartTime1()/60, menuOptions[activeMenu].getStartTime1()%60);
+    sprintf(displayText,"1 %02d%02d", (int)menuOptions[activeMenu].getStartTime1()/60, (int)menuOptions[activeMenu].getStartTime1()%60);
     if (changeDigit == MINUTES) blinkingSegment = 0b001100;
     else if (changeDigit == SECONDS) blinkingSegment = 0b000011;
     colonShown = true;
     break;
   case INTERVAL2:
-    sprintf(displayText,"2 %02d%02d", menuOptions[activeMenu].getStartTime2()/60, menuOptions[activeMenu].getStartTime2()%60);
+    sprintf(displayText,"2 %02d%02d", (int)menuOptions[activeMenu].getStartTime2()/60, (int)menuOptions[activeMenu].getStartTime2()%60);
     if (changeDigit == MINUTES) blinkingSegment = 0b001100;
     else if (changeDigit == SECONDS) blinkingSegment = 0b000011;
     colonShown = true;
